backend/requirements.txt
argcomplete==3.6.2
blinker==1.9.0
boto3==1.38.18
botocore==1.38.18
certifi==2025.4.26
cfn-flip==1.3.0
charset-normalizer==3.4.2
click==8.2.0
dnspython==2.7.0
durationpy==0.10
Flask==3.1.1
flask-cors==6.0.0
hjson==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==3.0.2
placebo==0.9.0
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.2
urllib3==2.4.0
Werkzeug==3.1.3
zappa==0.60.0
